refactor: PlayerResult의 속성으로 배팅 금액에 얼마가 곱해지는지만 나타내는 방법으로 수정

+++ src/main/java/blackjack/controller/BlackjackController.java
@@ -104,7 +104,7 @@ public class BlackjackController {
 
     private void calculateGameResultPrint(Dealer dealer, Players players) {
         PlayersProfit playersProfit = players.calculateProfit(dealer);
-        long dealerProfit = dealer.calculateProfit(playersProfit.allProfit());
+        double dealerProfit = dealer.calculateProfit(playersProfit.allProfit());
 
         OutputView.printFinalProfitAnnounce();
         OutputView.printDealerProfit(dealerProfit);
